+++ src/eca/features/chat.clj
@@ -52,6 +52,10 @@
   (:origin (first (filter #(= name (:name %)) all-tools))))
 
 (defn ^:private usage-msg->usage
+  "How this works:
+    - tokens: the last message from API already contain the total
+              tokens considered, but we save them for cost calculation
+    - cost: we count the tokens in past requests done + current one"
   [{:keys [input-tokens output-tokens
            input-cache-creation-tokens input-cache-read-tokens]}
    full-model
@@ -64,18 +68,17 @@
     (when input-cache-read-tokens
       (swap! db* update-in [:chats chat-id :total-input-cache-read-tokens] (fnil + 0) input-cache-read-tokens))
     (let [db @db*
-          message-input-cache-tokens (or input-cache-creation-tokens 0)
           total-input-tokens (get-in db [:chats chat-id :total-input-tokens] 0)
           total-input-cache-creation-tokens (get-in db [:chats chat-id :total-input-cache-creation-tokens] nil)
           total-input-cache-read-tokens (get-in db [:chats chat-id :total-input-cache-read-tokens] nil)
-          total-input-cache-tokens (or total-input-cache-read-tokens 0)
           total-output-tokens (get-in db [:chats chat-id :total-output-tokens] 0)
           model-capabilities (get-in db [:models full-model])]
-      (assoc-some {:message-output-tokens output-tokens
-                   :message-input-tokens (+ input-tokens message-input-cache-tokens)
-                   :session-tokens (+ total-input-tokens total-input-cache-tokens total-output-tokens)}
+      (assoc-some {:session-tokens (+ input-tokens
+                                      (or input-cache-read-tokens 0)
+                                      (or input-cache-creation-tokens 0)
+                                      output-tokens)}
                   :limit (:limit model-capabilities)
-                  :message-cost (shared/tokens->cost input-tokens input-cache-creation-tokens input-cache-read-tokens output-tokens model-capabilities)
+                  :last-message-cost (shared/tokens->cost input-tokens input-cache-creation-tokens input-cache-read-tokens output-tokens model-capabilities)
                   :session-cost (shared/tokens->cost total-input-tokens total-input-cache-creation-tokens total-input-cache-read-tokens total-output-tokens model-capabilities)))))
 
 (defn ^:private tokenize-args [^String s]